+++ src/pipe_log.rs
@@ -146,23 +146,45 @@ impl LogFileContext {
     }
 }
 
+/// Some bytes whose value might be dependent on the file it is written to.
+pub trait ReactiveBytes {
+    fn as_bytes(&mut self, ctx: &LogFileContext) -> &[u8];
+}
+
+impl<T> ReactiveBytes for &T
+where
+    T: AsRef<[u8]> + ?Sized,
+{
+    fn as_bytes(&mut self, _ctx: &LogFileContext) -> &[u8] {
+        (*self).as_ref()
+    }
+}
+
 /// A `PipeLog` serves reads and writes over multiple queues of log files.
+///
+/// # Safety
+///
+/// The pipe will panic if it encounters an unrecoverable failure. Otherwise the
+/// operations on it should be atomic, i.e. failed operation will not affect
+/// other ones, and user can still use it afterwards without breaking
+/// consistency.
 pub trait PipeLog: Sized {
     /// Reads some bytes from the specified position.
     fn read_bytes(&self, handle: FileBlockHandle) -> Result<Vec<u8>>;
 
     /// Appends some bytes to the specified log queue. Returns file position of
     /// the written bytes.
-    ///
-    /// The result of `fetch_active_file` will not be affected by this method.
-    fn append(&self, queue: LogQueue, bytes: &[u8]) -> Result<FileBlockHandle>;
+    fn append<T: ReactiveBytes + ?Sized>(
+        &self,
+        queue: LogQueue,
+        bytes: &mut T,
+    ) -> Result<FileBlockHandle>;
 
-    /// Hints it to synchronize buffered writes. The synchronization is
-    /// mandotory when `sync` is true.
+    /// Synchronizes all buffered writes.
     ///
     /// This operation might incurs a great latency overhead. It's advised to
     /// call it once every batch of writes.
-    fn maybe_sync(&self, queue: LogQueue, sync: bool) -> Result<()>;
+    fn sync(&self, queue: LogQueue) -> Result<()>;
 
     /// Returns the smallest and largest file sequence number, still in use,
     /// of the specified log queue.
@@ -194,8 +216,4 @@ pub trait PipeLog: Sized {
     ///
     /// Returns the number of deleted files.
     fn purge_to(&self, file_id: FileId) -> Result<usize>;
-
-    /// Returns [`LogFileContext`] of the active file in the specific
-    /// log queue.
-    fn fetch_active_file(&self, queue: LogQueue) -> LogFileContext;
 }